refactor: add :unit:replies module
