+++ domain/identity/build.gradle.kts
@@ -29,6 +29,8 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(libs.koin.core)
+                implementation(compose.foundation)
+                implementation(compose.runtime)
                 implementation(projects.core.preferences)
                 implementation(projects.core.api)
                 implementation(projects.core.utils)
