+++ settings.gradle.kts
@@ -77,3 +77,4 @@ include(":unit:selectcommunity")
 include(":unit:userdetail")
 include(":unit:web")
 include(":unit:zoomableimage")
+include(":unit:replies")
