+++ unit/replies/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/replies/di/Utils.kt
@@ -1,0 +1,9 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.replies.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.replies.InboxRepliesMviModel
+import org.koin.java.KoinJavaComponent
+
+actual fun getInboxRepliesViewModel(): InboxRepliesMviModel {
+    val res: InboxRepliesMviModel by KoinJavaComponent.inject(InboxRepliesMviModel::class.java)
+    return res
+}
