+++ unit/replies/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/replies/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.replies.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.replies.InboxRepliesMviModel
+
+expect fun getInboxRepliesViewModel(): InboxRepliesMviModel
